pyproject.toml
[project]
name = "01-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "chromadb>=1.1.0",
   "crewai>=0.201.1",
   "google-genai>=1.41.0",
   "markitdown[all]>=0.1.3",
   "python-dotenv>=1.1.1",
   "sentence-transformers>=5.1.1",
   "streamlit>=1.50.0",
]
